backend/requirements.txt
apispec==6.8.2
asgiref==3.9.1
decouple==0.0.7
Django==5.2.5
django-serializer==1.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
marshmallow==4.0.0
packaging==25.0
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
gunicorn
psycopg2-binary
whitenoise
psutil==5.9.8
amqp==5.3.1
apispec==6.8.2
asgiref==3.9.1
billiard==4.2.1
boto3==1.40.17
botocore==1.40.17
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
decouple==0.0.7
dj-database-url==3.0.1
Django==5.2.5
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-redis==6.0.0
django-serializer==1.2.1
django-storages==1.14.6
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
gunicorn==23.0.0
jmespath==1.0.1
kombu==5.5.4
marshmallow==4.0.0
packaging==25.0
prompt_toolkit==3.0.52
psutil==5.9.8
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.10.0
